benchmarks/rdi/requirements.txt
numpy
Pillow
scikit-image

# Optional
torch
lpips
